release 003 and 004

+++ src/main/groovy/com/daacs/model/item/ItemGroupTransition.groovy
@@ -15,5 +15,5 @@ class ItemGroupTransition {
     Difficulty groupDifficulty;
 
     @NotNull
-    Map<Difficulty, Range<Integer>> transitionMap;
+    Map<Difficulty, Range<Double>> transitionMap;
 }
